requirements.txt
aiofiles==0.8.0
aiogram==3.0.0b4
aiohttp==3.8.1
aioliqpay==1.0.0
aioliqpay-sdk-python3==1.0.0
aiomysql==0.1.1
aioschedule==0.5.2
aiosignal==1.2.0
aiosqlite==0.17.0
aiounittest==1.4.2
alembic==1.7.7
altair==4.2.0
altair-data-server==0.4.1
altair-viewer==0.4.0
anyio==3.6.1
appdirs==1.4.4
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
asyncpg==0.25.0
atomicwrites==1.4.1
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
bleach==5.0.1
boto3==1.24.16
botocore==1.27.16
Brotli==1.0.9
build==0.8.0
cachetools==5.2.0
capmonster-python==2.2
cattrs==1.10.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
cloudflare==2.9.10
cloudipsp==1.0.3
colorama==0.4.4
commonmark==0.9.1
cryptography==37.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
Deprecated==1.2.13
discord.py==2.0.0
discum==1.4.0
Django==4.0.4
django-jazzmin==2.5.0
dnspython==2.2.1
docutils==0.19
entrypoints==0.4
et-xmlfile==1.1.0
etherscan==0.3.0
etherscan-python==2.1.0
executing==0.8.3
ez-setup==0.9
Faker==13.12.0
fastapi==0.78.0
filetype==1.0.13
fonttools==4.34.4
frozenlist==1.3.0
geoip2==4.6.0
gforms==1.2.7
google-api-core==2.8.2
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
googleform==0.0.1a2
greenlet==1.1.2
h11==0.12.0
howdyho==3.0
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
install==1.3.5
ipykernel==6.13.0
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
jsonlines==3.0.0
jsonschema==4.7.2
jupyter-client==7.3.1
jupyter-core==4.10.0
keyring==23.6.0
kiwisolver==1.4.3
liqpay-sdk-python3==1.0.3
loguru==0.6.0
lxml==4.9.1
magic-filter==1.0.8
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
maxminddb==2.2.0
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.23.0
oauth2client==4.1.3
oauthlib==3.2.0
openpyxl==3.0.10
OPi.GPIO==0.5.2
packaging==21.3
pandas==1.4.3
paramiko==2.11.0
parso==0.8.3
pep517==0.12.0
pickleshare==0.7.5
Pillow==9.2.0
pip-tools==6.8.0
pkginfo==1.8.3
playwright==1.22.0
pluggy==1.0.0
portpicker==1.5.2
prompt-toolkit==3.0.29
protobuf==4.21.5
psutil==5.9.1
psycopg==3.0.15
psycopg-binary==3.0.15
psycopg-pool==3.1.1
psycopg2==2.9.3
pure-eval==0.2.2
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.2
pyee==8.1.0
Pygments==2.12.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
Pyrogram==2.0.26
pyrsistent==0.18.1
PySocks==1.7.1
pyTelegramBotAPI==4.5.1
pytest==7.1.2
pytest-asyncio==0.19.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytoolconfig==1.2.2
pytz==2022.1
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==23.1.0
readme-renderer==35.0
redis==4.3.4
requests==2.27.1
requests-cache==0.5.2
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==12.4.4
rope==1.3.0
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqladmin==0.1.12
SQLAlchemy==1.4.37
sqlparse==0.4.2
stack-data==0.2.0
starlette==0.19.1
telegraph==2.1.0
TgCrypto==1.2.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
twine==4.0.1
typing_extensions==4.2.0
tzdata==2022.1
ua-parser==0.10.0
ujson==5.4.0
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.9
uvicorn==0.17.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.59.0
websockets==10.1
win32-setctime==1.1.0
wrapt==1.14.1
WTForms==3.0.1
yarl==1.7.2
zipp==3.8.0
aiofiles==0.8.0
aiogram==3.0.0b4
aiohttp==3.8.1
aioliqpay==1.0.0
aioliqpay-sdk-python3==1.0.0
aiomysql==0.1.1
aioschedule==0.5.2
aiosignal==1.2.0
aiosqlite==0.17.0
aiounittest==1.4.2
alembic==1.7.7
altair==4.2.0
altair-data-server==0.4.1
altair-viewer==0.4.0
anyio==3.6.1
appdirs==1.4.4
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
asyncpg==0.25.0
atomicwrites==1.4.1
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
bleach==5.0.1
boto3==1.24.16
botocore==1.27.16
Brotli==1.0.9
build==0.8.0
cachetools==5.2.0
capmonster-python==2.2
cattrs==1.10.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
cloudflare==2.9.10
cloudipsp==1.0.3
colorama==0.4.4
commonmark==0.9.1
cryptography==37.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
Deprecated==1.2.13
discord.py==2.0.0
discum==1.4.0
Django==4.0.4
django-jazzmin==2.5.0
dnspython==2.2.1
docutils==0.19
entrypoints==0.4
et-xmlfile==1.1.0
etherscan==0.3.0
etherscan-python==2.1.0
executing==0.8.3
ez-setup==0.9
Faker==13.12.0
fastapi==0.78.0
filetype==1.0.13
fonttools==4.34.4
frozenlist==1.3.0
geoip2==4.6.0
gforms==1.2.7
google-api-core==2.8.2
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
googleform==0.0.1a2
greenlet==1.1.2
h11==0.12.0
howdyho==3.0
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
install==1.3.5
ipykernel==6.13.0
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
jsonlines==3.0.0
jsonschema==4.7.2
jupyter-client==7.3.1
jupyter-core==4.10.0
keyring==23.6.0
kiwisolver==1.4.3
liqpay-sdk-python3==1.0.3
loguru==0.6.0
lxml==4.9.1
magic-filter==1.0.8
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
maxminddb==2.2.0
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.23.0
oauth2client==4.1.3
oauthlib==3.2.0
openpyxl==3.0.10
OPi.GPIO==0.5.2
packaging==21.3
pandas==1.4.3
paramiko==2.11.0
parso==0.8.3
pep517==0.12.0
pickleshare==0.7.5
Pillow==9.2.0
pip-tools==6.8.0
pkginfo==1.8.3
playwright==1.22.0
pluggy==1.0.0
portpicker==1.5.2
prompt-toolkit==3.0.29
protobuf==4.21.5
psutil==5.9.1
psycopg==3.0.15
psycopg-binary==3.0.15
psycopg-pool==3.1.1
psycopg2==2.9.3
pure-eval==0.2.2
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.2
pyee==8.1.0
Pygments==2.12.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
Pyrogram==2.0.26
pyrsistent==0.18.1
PySocks==1.7.1
pyTelegramBotAPI==4.5.1
pytest==7.1.2
pytest-asyncio==0.19.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytoolconfig==1.2.2
pytz==2022.1
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==23.1.0
readme-renderer==35.0
redis==4.3.4
requests==2.27.1
requests-cache==0.5.2
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==12.4.4
rope==1.3.0
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqladmin==0.1.12
SQLAlchemy==1.4.37
sqlparse==0.4.2
stack-data==0.2.0
starlette==0.19.1
telegraph==2.1.0
TgCrypto==1.2.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
twine==4.0.1
typing_extensions==4.2.0
tzdata==2022.1
ua-parser==0.10.0
ujson==5.4.0
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.9
uvicorn==0.17.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.59.0
websockets==10.1
win32-setctime==1.1.0
wrapt==1.14.1
WTForms==3.0.1
yarl==1.7.2
zipp==3.8.0
